+++ src/commands/index.js
@@ -1,0 +1,5 @@
+const runTemplator = require('./run-templator')
+
+module.exports = {
+  runTemplator
+}