+++ src/utils/index.js
@@ -1,0 +1,5 @@
+const getDefaultPromptMessage = variableName => `Enter value for ${variableName}`
+
+module.exports = {
+  getDefaultPromptMessage
+}